Structural Issue: The console consumes * characters, preventing simple math functions - FIXED
Structural Issue: The console consumes * characters, preventing simple math functions - FIXED

The * character is used for diagnostic queries between the machine and the controller. It is not used by smoothieware and is a makera addition.

When a * is sent through the console to the machine, the machine immediately sends the diagnostic string back and removes it from the serial queue. This means it never reaches the gcode interpreter.

By consuming * characters, the firmware prevents us from implementing multiplication as 1*1 returns 2 in gcode, but 11 in the console. (in our dev branch),

If the character were changed to '&' instead, there would be no conflicts.

This would also require a change to controller.py
line 718 self.stream.send(b"*") -> self.stream.send(b"&") and 397 change the * to &
39acb8a
src\modules\communication\SerialConsole.cpp